Moved all shift methods from cracker into a ShiftCalculator module
